mcom02_flash_ums_mmc: Use wait_for_uboot() for determine U-Boot
During the countdown, U-Boot only updates the seconds value. So if
the script is run after the output "Hit any key ... ", but during
the countdown, the script will not determine that the countdown is
running.

Function wait_fot_uboot() is now used to determine U-Boot loading.
This function correctly handle the following situation:

  * when U-Boot countdown is already running;
  * when U-Boot already loaded;
  * waiting for U-Boot "Hit any key to stop autoboot..." message.

Change-Id: I4e4322f2e29866feb22ea069c4ef06f33ac3fe9f